lepiota/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.10.0
django-oauth-toolkit==1.3.3
djangorestframework==3.12.1
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
